+++ spyder/app/tests/test_mainwindow.py
@@ -146,17 +146,6 @@ def find_desired_tab_in_window(tab_name, window):
 # =============================================================================
 # ---- Fixtures
 # =============================================================================
-@pytest.hookimpl(tryfirst=True, hookwrapper=True)
-def pytest_runtest_makereport(item, call):
-    # execute all other hooks to obtain the report object
-    outcome = yield
-    rep = outcome.get_result()
-
-    # set a report attribute for each phase of a call, which can
-    # be "setup", "call", "teardown"
-
-    setattr(item, "rep_" + rep.when, rep)
-
 
 @pytest.fixture
 def main_window(request):
